requirements.txt
aiofiles
aioshutil
Pillow
PyYAML
regex
rustworkx
tqdm
wcmatch
